Separated UserTitlesService class
